Add files via upload
The notebook can generate smooth transmission line shapes in the coupler, with the start and the end connected to the fixed feeder line patterns. Shapes are visualized in every step, and the current notebook shows the pattern evolution in the 3rd iteration. Only the S3-S4 transmission line is optimized.